+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### v0.6dev
 * New [Salmon](http://combine-lab.github.io/salmon/) module.
 * New [Trimmomatic](http://www.usadellab.org/cms/?page=trimmomatic) module.
+* New [Bamtools stats](https://github.com/pezmaster31/bamtools) module.
 * New beeswarm plot type. General Stats table replaced with this when many samples in report.
 * New RSeQC module: Actually a suite of 8 new modules supporting various outputs from RSeQC
 * Rewrote bowtie2 module: Now better at parsing logs and tries to scrape input from wrapper logs.